+++ apps/web/src/features/call/in-call/hooks/useInCallListeners.ts
@@ -24,14 +24,4 @@ export const useInCallListeners = (socket: Maybe<VChatSocket>) => {
       socket?.off('partner-disconnected', onPartnerDisconnected);
     };
   }, [onPartnerDisconnected, onUserLeft, socket]);
-
-  useEffect(() => {
-    const handleBeforeUnload = () => {
-      InCallService.dispose();
-    };
-
-    return () => {
-      window.addEventListener('beforeunload', handleBeforeUnload);
-    };
-  }, []);
 };
