+++ apps/web/src/features/call/in-call/page/__tests__/InCallPage.test.tsx
@@ -1,7 +1,8 @@
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import { render, screen } from '@testing-library/react';
+import { RoutePath } from '@/RoutePath';
 import { CallRouterStateLocation, InCallPage } from '../InCallPage';
-import { RoutePath } from '../../../../../RoutePath';
+
 vi.mock('react-router-dom', () => ({
   useLocation: vi.fn(),
   useNavigate: vi.fn(),
@@ -42,7 +43,10 @@ describe('InCallPage', () => {
 
     render(<InCallPage />);
 
-    expect(mockNavigate).toHaveBeenCalledWith(RoutePath.CALL, { state: null });
+    expect(mockNavigate).toHaveBeenCalledWith(RoutePath.HOME, {
+      state: null,
+      replace: true,
+    });
   });
 
   it('should navigate away when state is missing', () => {
@@ -51,7 +55,10 @@ describe('InCallPage', () => {
 
     render(<InCallPage />);
 
-    expect(mockNavigate).toHaveBeenCalledWith(RoutePath.CALL, { state: null });
+    expect(mockNavigate).toHaveBeenCalledWith(RoutePath.HOME, {
+      state: null,
+      replace: true,
+    });
   });
 
   it('should render call container when both roomId and state are valid', () => {
@@ -83,6 +90,9 @@ describe('InCallPage', () => {
     const { container } = render(<InCallPage />);
 
     expect(container.firstChild).toBeNull();
-    expect(mockNavigate).toHaveBeenCalledWith(RoutePath.CALL, { state: null });
+    expect(mockNavigate).toHaveBeenCalledWith(RoutePath.HOME, {
+      state: null,
+      replace: true,
+    });
   });
 });
