--- apps/web/src/features/call/queue/component/Video.tsx
@@ -1,43 +1,0 @@
-import { BsCameraVideoOff } from 'react-icons/bs';
-import { TypographyP } from '@/common/components/typography/Typography';
-
-interface Props {
-  videoRef?: React.RefObject<HTMLVideoElement>;
-  isLocal?: boolean;
-  videoEnabled: boolean;
-  isPortrait: boolean;
-}
-
-export function Video({ videoRef, isLocal, videoEnabled, isPortrait }: Props) {
-  return (
-    <div className="relative flex justify-center w-full h-full">
-      <video
-        ref={videoRef}
-        className={`${
-          isPortrait ? 'w-auto h-full' : 'w-full h-auto'
-        } ${isLocal ? 'z-10' : 'z-0'}`}
-        autoPlay
-        playsInline
-        muted={isLocal}
-      />
-
-      {!videoEnabled && (
-        <div
-          className={`absolute inset-0 flex items-center justify-center ${isLocal ? 'z-10 shadow-video-off' : 'z-0'}`}
-        >
-          <BsCameraVideoOff
-            className={`text-white ${isLocal ? 'text-xl' : 'text-4xl'}`}
-            aria-label={
-              isLocal ? 'Your camera is off' : "Partner's camera is off icon"
-            }
-          />
-          {!isLocal && (
-            <TypographyP className="ml-2 text-white" noFirstMarginTop>
-              Partner's camera is off
-            </TypographyP>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
