--- apps/web/src/features/call/queue/container/UserVideoContainer.tsx
@@ -1,14 +1,0 @@
-import { useIsPortraitVideoOrientation } from '@/common/hooks/useIsPortraitVideoOrientation';
-import { Video } from '../component/Video';
-
-interface Props {
-  videoRef?: React.RefObject<HTMLVideoElement>;
-  isLocal?: boolean;
-  videoEnabled: boolean;
-}
-
-export function UserVideoContainer({ videoRef, ...props }: Props) {
-  const isPortrait = useIsPortraitVideoOrientation(videoRef);
-
-  return <Video videoRef={videoRef} {...props} isPortrait={isPortrait} />;
-}
