+++ apps/web/src/features/game/game-engine/model/GameState.ts
@@ -1,0 +1,9 @@
+export enum GameState {
+  IDLE = 'IDLE',
+  PREPARE_ROUND = 'PREPARE_ROUND',
+  ROUND_START = 'ROUND_START',
+  PLAYER_TURN = 'PLAYER_TURN',
+  SPECTATOR_TURN = 'SPECTATOR_TURN',
+  ROUND_END = 'ROUND_END',
+  GAME_OVER = 'GAME_OVER',
+}
