+++ apps/web/src/features/game/game-invite/component/GameInviteButton.tsx
@@ -1,3 +1,4 @@
+import { observer } from 'mobx-react';
 import { Gamepad2 } from 'lucide-react';
 import { Button } from '@/common/components/ui/button';
 
@@ -6,7 +7,10 @@ interface Props {
   disabled?: boolean;
 }
 
-export function GameInviteButton({ onClick, disabled }: Props) {
+export const GameInviteButton = observer(function GameInviteButton({
+  onClick,
+  disabled,
+}: Props) {
   return (
     <Button
       aria-label="Invite to game"
@@ -18,4 +22,4 @@ export function GameInviteButton({ onClick, disabled }: Props) {
       <Gamepad2 className="h-6 w-6" />
     </Button>
   );
-}
+});
